+++ openfisca_france_data/tests/erfs_fpr/test_af.py
@@ -5,28 +5,10 @@
 from __future__ import division
 
 
-import numpy as np
+from openfisca_france_data.erfs_fpr.get_survey_scenario import get_survey_scenario
 
 
-from openfisca_france_data.erfs_fpr.scenario import ErfsFprSurveyScenario
-from openfisca_france_data.tests import base as base_survey
-
-
-def get_survey_scenario(year = 2012, rebuild_input_data = False):
-    tax_benefit_system = base_survey.get_cached_reform(
-        reform_key = 'inversion_directe_salaires',
-        tax_benefit_system = base_survey.france_data_tax_benefit_system,
-        )
-    survey_scenario = ErfsFprSurveyScenario.create(
-        tax_benefit_system = tax_benefit_system,
-        baseline_tax_benefit_system = tax_benefit_system,
-        year = year,
-        rebuild_input_data = rebuild_input_data,
-        )
-    return survey_scenario
-
-
-survey_scenario = get_survey_scenario()
+survey_scenario = get_survey_scenario(year = 2012, reform_key = 'inversion_directe_salaires')
 
 #%%
 data_frame_by_entity = survey_scenario.create_data_frame_by_entity(
@@ -84,7 +66,9 @@ menage.groupby(['residence_dom'])['wprm'].sum()
 individu.groupby(['prestations_familiales_enfant_a_charge'])['weight_individus'].sum()  # PROBLEM
 individu.groupby(['est_enfant_dans_famille'])['weight_individus'].sum()
 individu.groupby(['autonomie_financiere'])['weight_individus'].sum()
-individu.groupby(['autonomie_financiere', 'est_enfant_dans_famille', 'rempli_obligation_scolaire'])['weight_individus'].sum()
+individu.groupby(
+    ['autonomie_financiere', 'est_enfant_dans_famille', 'rempli_obligation_scolaire']
+    )['weight_individus'].sum()
 
 #%%
 survey_scenario.get_memory_usage('rempli_obligation_scolaire')