+++ openfisca_france_data/tests/erfs_fpr/test_al.py
@@ -7,25 +7,10 @@
 from __future__ import division
 
 
-from openfisca_france_data.erfs_fpr.scenario import ErfsFprSurveyScenario
-from openfisca_france_data.tests import base as base_survey
+from openfisca_france_data.erfs_fpr.get_survey_scenario import get_survey_scenario
 
 
-def get_survey_scenario(year = 2012, rebuild_input_data = False):
-    tax_benefit_system = base_survey.get_cached_reform(
-        reform_key = 'inversion_directe_salaires',
-        tax_benefit_system = base_survey.france_data_tax_benefit_system,
-        )
-    survey_scenario = ErfsFprSurveyScenario.create(
-        tax_benefit_system = tax_benefit_system,
-        baseline_tax_benefit_system = tax_benefit_system,
-        year = year,
-        rebuild_input_data = rebuild_input_data,
-        )
-    return survey_scenario
-
-
-survey_scenario = get_survey_scenario()
+survey_scenario = get_survey_scenario(year = 2012, reform_key = 'inversion_directe_salaires')
 
 #%%
 data_frame_by_entity = survey_scenario.create_data_frame_by_entity(
@@ -59,6 +44,9 @@ assert .39 < statut_occupation_logement_pct[3:7].sum() < .43
 #%%
 
 11e10 < (famille.aide_logement_montant_brut * famille.weight_familles).sum()
-11e10 < (famille.apl * famille.weight_familles).sum() + (famille.als * famille.weight_familles).sum() + (famille.alf * famille.weight_familles).sum()
-
+11e10 < (
+    (famille.apl * famille.weight_familles).sum()
+    + (famille.als * famille.weight_familles).sum()
+    + (famille.alf * famille.weight_familles).sum()
+    )
 ## Reste un peu faiblard