+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.6
+
+* Migration vers la version 23 d'openfisca-core
+
+## 0.5.10
+
+* Migration vers la version 21 d'openfisca-core
+
+
 ## 0.5.7
 
 * Migration vers la nouvelle syntaxe (abandon des entity_key_plural)